+++ R/outliers.R
@@ -9,7 +9,7 @@ NULL
 setMethod(
   f = "outliers",
   signature = c(object = "CompositionMatrix"),
-  definition = function(object, ..., groups = get_groups(object), robust = TRUE,
+  definition = function(object, ..., groups = get_groups(object),
                         method = c("mve", "mcd"), quantile = 0.975) {
     ## Transformation
     z <- transform_ilr(object)
@@ -19,53 +19,60 @@ setMethod(
     p <- ncol(z)
     if (is.null(groups) || all(is.na(groups))) {
       grp <- list(z)
-      groups <- rep(NA_character_, m)
+      groups <- list(seq_len(m))
     } else {
       grp <- split(z, f = groups)
+      groups <- split(seq_len(m), f = groups)
     }
 
     ## Clean
     size <- vapply(X = grp, FUN = nrow, FUN.VALUE = integer(1), USE.NAMES = TRUE)
-    too_small <- size < p
+    too_small <- size < (p + 1)
+    very_small <- size < (2 * p)
+    if (all(too_small)) {
+      stop("Too few samples.", call. = FALSE)
+    }
     if (any(too_small)) {
       msg <- "%s is ignored: sample size is too small (%d).\n"
       warning(sprintf(msg, names(grp)[too_small], size[too_small]), call. = FALSE)
     }
-    if (all(too_small)) {
-      stop("Too few samples.", call. = FALSE)
+    if (any(very_small)) {
+      msg <- "%s: possibly too small sample size (%d).\n"
+      warning(sprintf(msg, names(grp)[very_small], size[very_small]), call. = FALSE)
     }
-    grp <- grp[!too_small]
 
-    d2 <- matrix(data = NA_real_, nrow = m, ncol = length(grp))
-    dimnames(d2) <- list(rownames(object), names(grp))
+    dc <- dr <- matrix(data = NA_real_, nrow = m, ncol = length(grp),
+                       dimnames = list(rownames(object), names(grp)))
 
     ## For each group...
-    for (i in seq_along(grp)) {
+    for (i in seq_along(grp)[!too_small]) {
       ## ...subset
       tmp <- grp[[i]]
 
       ## ...compute center and spread
-      if (!robust) method <- "classical" # Standard estimators
-      else method <- match.arg(method, several.ok = FALSE) # Robust estimators
-      est <- MASS::cov.rob(tmp, method = method, ...)
+      ## Standard estimators
+      estc <- list(center = colMeans(tmp, na.rm = TRUE), cov = cov(tmp))
 
-      ## Mahalanobis distance
-      d2[, i] <- stats::mahalanobis(z, center = est$center, cov = est$cov)
+      ## Robust estimators
+      method <- match.arg(method, several.ok = FALSE)
+      estr <- MASS::cov.rob(tmp, method = method, ...)
+
+      ## ...Mahalanobis distance
+      dc[, i] <- stats::mahalanobis(z, center = estc$center, cov = estc$cov)
+      dr[, i] <- stats::mahalanobis(z, center = estr$center, cov = estr$cov)
     }
 
     ## Threshold
-    dof <- p - 1L
-    limit <- sqrt(stats::qchisq(p = quantile, df = dof))
-
-    d <- sqrt(d2)
+    limit <- sqrt(stats::qchisq(p = quantile, df = p))
 
     .OutlierIndex(
-      d,
       samples = get_samples(object),
-      groups = as.character(groups),
+      groups = groups,
+
+      standard = sqrt(dc),
+      robust = sqrt(dr),
       limit = limit,
-      robust = robust,
-      dof = dof
+      dof = p
     )
   }
 )